// src/Moryx.Operators.Web/src/app/api/models/skill-model.ts
/*
 * Copyright (c) 2026 Phoenix Contact GmbH & Co. KG
 * Licensed under the Apache License, Version 2.0
*/

/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface SkillModel {
  expiresOn?: string;
  id?: number;
  isExpired?: boolean;
  obtainedOn?: string;
  operatorIdentifier?: string | null;
  typeId?: number;
}