// src/Moryx.Operators.Web/src/app/api/models/skill-type-model.ts
/*
 * Copyright (c) 2026 Phoenix Contact GmbH & Co. KG
 * Licensed under the Apache License, Version 2.0
*/

/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { Entry } from '@moryx/ngx-web-framework/entry-editor';
export interface SkillTypeModel {
  capabilities?: Entry;
  duration?: string;
  id?: number;
  name?: string | null;
}